Attempt to move recording
