+++ src/session-recording/loader-recorder-v2.ts
