+++ src/session-recording/types.ts
@@ -1,0 +1,1 @@
+export interface SesionRecorder {}
